Rename icon filter nav svgs
