+++ playbook-website/app/javascript/images/kit_previews/navigation/index.ts
@@ -1,5 +1,5 @@
 import breadCrumbs from "./bread_crumbs.svg";
-import nav from "./nav.svg";
+import nav from "./kit_nav.svg";
 
 export const navigationPreviewMap = {
   "navigation/bread_crumbs": breadCrumbs,
